% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TopDomData-package}
\alias{TopDomData-package}
\alias{TopDomData}
\title{TopDomData - Data for the TopDom Package}
\description{
The \pkg{TopDomData} package is an \R package that provides examples data
for the \pkg{TopDom} package.  This data package is also used for
validation of implementation of the TopDom package via TopDom's package
tests.
}
\section{Origin}{

The data herein contain a tiny subset of the HiC and TopDom data obtained
via \href{http://zhoulab.usc.edu/TopDom/}{TopDom website}.
More precisely, it contains:
\enumerate{
\item A TopDom file \file{mESC_5w_chr10.nij.HindIII.comb.40kb.domain}, which
is part of the \file{mESC_5w_domain.zip} file
(54,676 bytes; md5 1ebfb9a556c1954b8cd35520a2499abd) from the
Supplementary Materials of Shin et al. (2015).
These data were downloaded from the
\href{http://zhoulab.usc.edu/TopDom/}{TopDom website}.
\item A normalized HiC-count matrix file \file{nij.chr10.gz}, where the
non-compressed version is part of the \file{mESC.norm.tar.gz} file
(1,305,763,679 bytes; md5 2e79d0f57463b5b7c4bf86b187086d3c) originally
downloaded from the
\href{http://chromosome.sdsc.edu/mouse/hi-c/download.html}{UCSD Ren Lab}.
}
}

\section{License}{

The license for these data is unknown.
}

\references{
Shin et al., TopDom: an efficient and deterministic method for identifying
topological domains in genomes, Nucleic Acids Res. 2016 Apr 20;
44(7): e70., 2015. doi: 10.1093/nar/gkv1505,
PMCID: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4838359/}{PMC4838359},
PMID: 26704975.
}
