% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TopDomData-package}
\alias{TopDomData-package}
\alias{TopDomData}
\title{TopDomData - Data for the TopDom Package}
\description{
The \pkg{TopDomData} package is an \R package that provides examples data
for the \pkg{TopDom} package.  This data package is also used for
validation of implementation of the TopDom package via TopDom's package
tests.
}
\section{Origin}{

The data herein contain a tiny subset of the HiC and TopDom data used
in the TopDom study (Shin et al., 2016).
More precisely, it contains:
\enumerate{
\item A TopDom file \file{mESC_5w_chr10.nij.HindIII.comb.40kb.domain}, which
is part of the \file{mESC_5w_domain.zip} file
(54,676 bytes; md5 1ebfb9a556c1954b8cd35520a2499abd) from the
Supplementary Materials of Shin et al. (2016).
These data were downloaded from the
TopDom website (http://zhoulab.usc.edu/TopDom/ - now defunct).
\item A normalized HiC-count matrix file \file{nij.chr10.gz}, where the
non-compressed version is part of the \file{mESC.norm.tar.gz} file
(1,305,763,679 bytes; md5 2e79d0f57463b5b7c4bf86b187086d3c) originally
downloaded from the
\href{http://chromosome.sdsc.edu/mouse/hi-c/download.html}{UCSD Ren Lab}.
It is a tab-delimited file containing a 3250-by-3250 numeric matrix
non-negative decimal values. The underlying HiC sequence data is
available from
\href{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35156}{GSE35156}
on GEO and was published part of Dixon, et al. (2012).
}
}

\section{License}{

The license for these data is unknown.
}

\references{
\enumerate{
\item Dixon JR, Selvaraj S, Yue F, Kim A, et al. Topological domains in
mammalian genomes identified by analysis of chromatin interactions.
Nature 2012 Apr 11; 485(7398):376-80,
doi: 10.1038/nature11082,
PMCID: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3356448/}{PMC3356448},
PMID: 22495300.
\item Shin, et al., TopDom: an efficient and deterministic method for
identifying topological domains in genomes,
Nucleic Acids Res. 2016 Apr 20; 44(7): e70., 2016.
doi: 10.1093/nar/gkv1505,
PMCID: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4838359/}{PMC4838359},
PMID: 26704975.
}
}
